refactor: richtext editor to standalone component
